Пока в процессе, думаю как реализовать метод findAll на Orders. Структуру базы изменил.

+++ src/main/java/entity/Product.java
@@ -4,11 +4,11 @@ import java.math.BigDecimal;
 import java.util.Objects;
 
 public class Product {
-    private final Integer id;
+    private final Long id;
     private final String name;
     private final BigDecimal price;
 
-    public Product(Integer id, String name, BigDecimal price) {
+    public Product(Long id, String name, BigDecimal price) {
         this.id = id;
         this.name = name;
         this.price = price;
